update stock plugins
